requirements.txt
discord.py
eth-brownie